feat: rename plan to visualize
We intend to support viewing an already run deployment, hence plan makes
less sense. Instead we will go with the more generic visualize for the
moment.
